A new post-processing type for verifying hit questions using LLM has been added (#669)
I think that using LLM for verification can yield better results in some cases where quality requirements are higher. Especially in the customer service field and the like, other users' questions may lead to privacy leaks or mention other brands, thus causing interference and so on.

In terms of specific details, I maintained the operation of openai==0.28.0 on the interface and was compatible with the operation of openai>1.0.0 and above.

I have added the test case, the example test file, and updated the example/readme.md.

I upgraded the version of onnxruntime to 1.21.1 because the previous version 1.14.0 is no longer in use.